Fix failing tests, extract admin_client fixture
