+++ tests/database/test_delete.py
@@ -20,4 +20,5 @@ def test_does_not_exist(admin_client):
 
     assert result is False
 
+
 # TODO build out doc cleanup tests
