+++ tests/document/test_init.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, print_function, unicode_literals
 
-
 from pysyncgateway import Document
